cellos_kernel::task::pipe

Struct Pipe

struct Pipe {
    buffer: Vec<u8>,
    head: usize,
    len: usize,
    readers: usize,
    writers: usize,
    read_waiters: VecDeque<usize>,
    write_waiters: VecDeque<usize>,
    owner_cell: usize,
}

Fields
buffer: Vec<u8>head: usizelen: usizereaders: usizewriters: usizeread_waiters: VecDeque<usize>write_waiters: VecDeque<usize>owner_cell: usize

Cell charged for the ring; the charge is refunded when the pipe is dropped.

Implementations

impl Pipe

fn capacity(&self) -> usize

fn is_empty(&self) -> bool

fn is_full(&self) -> bool

fn write(&mut self, src: &[u8]) -> usize

Copy src into the ring, returning how many bytes fit.

fn read(&mut self, dst: &mut [u8]) -> usize

Copy up to dst.len() bytes out of the ring.